split integrations.rs into several modules and fmt+clippy applied
